+++ src/data/series/grouping/aggregation/FloatMax.java
@@ -9,7 +9,7 @@ public class FloatMax implements FloatAggregateFunction {
     @Override
     public float group(FloatSeries series, long from, int length) {
         float max = series.get(from);
-        for (int i = from + 1; i < from + length; i++) {
+        for (long i = from + 1; i < from + length; i++) {
             max = Math.max(max, series.get(i));
         }
         return max;
