+++ src/data/series/grouping/aggregation/IntMax.java
@@ -9,7 +9,7 @@ public class IntMax implements IntAggregateFunction {
     @Override
     public int group(IntSeries series, long from, int length) {
         int max = series.get(from);
-        for (int i = from + 1; i < from + length; i++) {
+        for (long i = from + 1; i < from + length; i++) {
             max = Math.max(max, series.get(i));
         }
         return max;
